% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{as.data.frame.Dataset}
\alias{as.data.frame.Dataset}
\title{'as.data.frame' method for \code{Dataset} objects}
\usage{
\method{as.data.frame}{Dataset}(x, row.names = NULL, optional = FALSE, ..., columns = "all")
}
\arguments{
\item{x}{A Dataset object}

\item{row.names}{A vector of row names}

\item{optional}{A logical value. If TRUE, setting row names and converting column names is optional.}

\item{...}{Additional arguments}

\item{columns}{Which columns to include in the data frame: "all", "target", "features". By default "all".}
}
\description{
Convert a Dataset object to a data frame
}
\examples{
cars.data <- Dataset(data = cars[c(1, 2, 3, 4), ], target = "dist")
as.data.frame(cars.data)
}
