% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{configuration<-}
\alias{configuration<-}
\title{Setter for the \code{configuration} of an inducer.}
\usage{
configuration(object) <- value

configuration(object) <- value
}
\arguments{
\item{object}{S3 object of class \code{Inducer}.}

\item{value}{Named list of configurations.}
}
\description{
S3 generic. Set new configuration values for the hyperparameters of a given inducer.
}
