% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{configuration.Inducer}
\alias{configuration.Inducer}
\title{Access \code{configuration} of an inducer.}
\usage{
\method{configuration}{Inducer}(x)
}
\arguments{
\item{x}{S3 object of class \code{Inducer}.}
}
\description{
Access \code{configuration} of an inducer.
}
