% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration.Model}
\alias{configuration.Model}
\title{Configuration Information}
\usage{
\method{configuration}{Model}(x)
}
\arguments{
\item{x}{A model object}
}
\description{
this functions returns the configuration of a model
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
model.xgb <- fit(xgb, cars.data)
configuration(model.xgb)
}
