% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{configuration}
\alias{configuration}
\title{Accessor for the \code{configuration} of an inducer.}
\usage{
configuration(x)
}
\arguments{
\item{x}{S3 object}
}
\value{
Named list of hyperparameter configurations.
}
\description{
S3 generic
}
