% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{configuration.Split}
\alias{configuration.Split}
\title{'configuration' method for \code{Split} objects}
\usage{
\method{configuration}{Split}(x)
}
\arguments{
\item{x}{A Split object}
}
\description{
Get the configuration of a Split object
}
\examples{
cv5 <- splt$cv(folds = 5)
configuration(cv5)
}
