% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{copy.Inducer}
\alias{copy.Inducer}
\title{Copy an inducer.}
\usage{
\method{copy}{Inducer}(.inducer, new_configuration)
}
\arguments{
\item{.inducer}{S3 object of class \code{Inducer}.}

\item{new_configuration}{Named List containing the hyperparameter}
}
\description{
Method that produces copies of instances of \code{Inducer} class that contain a
new hyperparameter configuration.
}
