% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Dataset S3 class}
\usage{
Dataset(data, target, name = deparse(substitute(data)), type)
}
\arguments{
\item{data}{A data frame containing the data}

\item{target}{The name of the target column}

\item{name}{The name of the dataset, by default the name of the given data object}

\item{type}{The task of the dataset: "classification" or "regression". If no task is provided, numerical target columns are considered as regression tasks, and non-numeric target columns as classification tasks.}
}
\description{
A class for storing datasets
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
class(cars.data)
}
