% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAccuracy.data.frame}
\alias{EvaluatorAccuracy.data.frame}
\title{Accuracy method for data.frame}
\usage{
\method{EvaluatorAccuracy}{data.frame}(.prediction)
}
\arguments{
\item{.prediction}{A data.frame where prediction column is the predicted class, and truth column is the actual class.}
}
\description{
Calculates Accuracy for a data.frame.
}
