% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAccuracy}
\alias{EvaluatorAccuracy}
\title{Accuracy generic function}
\usage{
EvaluatorAccuracy(.prediction)
}
\arguments{
\item{.prediction}{An object containing the predictions and the truth.}
}
\description{
This function calculates accuracy.
}
