% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAccuracy.ResamplePrediction}
\alias{EvaluatorAccuracy.ResamplePrediction}
\title{Accuracy method for ResamplePrediction}
\usage{
\method{EvaluatorAccuracy}{ResamplePrediction}(.prediction)
}
\arguments{
\item{.prediction}{A ResamplePrediction object.}
}
\description{
Calculates Accuracy for a ResamplePrediction object.
}
