% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAUC.data.frame}
\alias{EvaluatorAUC.data.frame}
\title{AUC method for data.frame}
\usage{
\method{EvaluatorAUC}{data.frame}(.prediction)
}
\arguments{
\item{.prediction}{A data.frame with prediction probabilities and actual binary outcomes.}
}
\description{
Calculates AUC for a data.frame.
}
