% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAUC}
\alias{EvaluatorAUC}
\title{AUC generic function}
\usage{
EvaluatorAUC(.prediction)
}
\arguments{
\item{.prediction}{An object containing the predictions and the truth.}
}
\description{
This function calculates the area under the ROC curve.
}
