% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorAUC.ResamplePrediction}
\alias{EvaluatorAUC.ResamplePrediction}
\title{AUC method for ResamplePrediction}
\usage{
\method{EvaluatorAUC}{ResamplePrediction}(.prediction)
}
\arguments{
\item{.prediction}{A ResamplePrediction object.}
}
\description{
Calculates AUC for a ResamplePrediction object.
}
