% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorMAE.data.frame}
\alias{EvaluatorMAE.data.frame}
\title{Mean Absolute Error method for data.frame}
\usage{
\method{EvaluatorMAE}{data.frame}(.prediction)
}
\arguments{
\item{.prediction}{A data.frame with two columns: prediction and truth}
}
\description{
This method calculates the mean absolute error for a data.frame
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
model.xgb <- fit(xgb, cars.data)
prediction <- predict(model.xgb, newdata = cars.data[c(1, 2, 3, 4), ])
EvaluatorMAE(prediction)
}
