% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorMAE}
\alias{EvaluatorMAE}
\title{Mean Absolute Error generic function}
\usage{
EvaluatorMAE(.prediction)
}
\arguments{
\item{.prediction}{An object containing the predictions and the truth}
}
\description{
This function calculates the mean absolute error
}
