% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorMAE.ResamplePrediction}
\alias{EvaluatorMAE.ResamplePrediction}
\title{Mean Absolute Error method for ResamplePrediction}
\usage{
\method{EvaluatorMAE}{ResamplePrediction}(.prediction)
}
\arguments{
\item{.prediction}{A ResamplePrediction object}
}
\description{
This method calculates the mean absolute error for a ResamplePrediction object
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
cv5 <- splt$cv(folds = 5)
rp <- resample(cars.data, xgb, cv5)
#EvaluatorMAE(rp)
}
