% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorRMSE.data.frame}
\alias{EvaluatorRMSE.data.frame}
\title{RMSE method for data.frame}
\usage{
\method{EvaluatorRMSE}{data.frame}(.prediction)
}
\arguments{
\item{.prediction}{A data.frame with two columns: prediction and truth.}
}
\description{
Calculates RMSE for a data.frame.
}
