% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorRMSE}
\alias{EvaluatorRMSE}
\title{Root Mean Squared Error generic function}
\usage{
EvaluatorRMSE(.prediction)
}
\arguments{
\item{.prediction}{An object containing the predictions and the truth.}
}
\description{
This function calculates the root mean squared error.
}
