% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{EvaluatorRMSE.ResamplePrediction}
\alias{EvaluatorRMSE.ResamplePrediction}
\title{RMSE method for ResamplePrediction}
\usage{
\method{EvaluatorRMSE}{ResamplePrediction}(.prediction)
}
\arguments{
\item{.prediction}{A ResamplePrediction object.}
}
\description{
Calculates RMSE for a ResamplePrediction object.
}
