% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\docType{data}
\name{evaluators_list}
\alias{evaluators_list}
\title{Evaluators Environment}
\format{
An object of class \code{list} of length 4.
}
\usage{
evaluators_list
}
\description{
Evaluators Environment
}
\keyword{datasets}
