% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{fit.Inducer}
\alias{fit.Inducer}
\title{Fit a model defined by an \code{Inducer} object on data defined by a \code{Dataset} object.}
\usage{
\method{fit}{Inducer}(.inducer, .data, ...)
}
\arguments{
\item{.inducer}{S3 object of class \code{Inducer}.}

\item{.data}{S3 object of class \code{Dataset}.}

\item{...}{Additional arguments.}
}
\description{
Fit a model defined by an \code{Inducer} object on data defined by a \code{Dataset} object.
}
