% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{fit}
\alias{fit}
\title{S3 generic for model fit on inducers.}
\usage{
fit(.inducer, .data, ...)
}
\arguments{
\item{.inducer}{S3 object of class Inducer.}

\item{.data}{S3 object of class Dataset.}

\item{...}{Additional arguments.}
}
\value{
S3 object of class \code{Model}.
}
\description{
S3 generic for model fit on inducers.
}
