% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerGlm.R
\name{fit2.InducerGlm}
\alias{fit2.InducerGlm}
\title{Fit a GLM.}
\usage{
\method{fit2}{InducerGlm}(.inducer, task, training_data, response)
}
\arguments{
\item{.inducer}{An S3 object of class \code{InducerGlm}.}

\item{task}{String that is either of value "Regression" or "Classification".}

\item{training_data}{Data.Frame or Matrix extracted from the \code{data} field
of the \code{Dataset} class.}

\item{response}{String that stores the name of the target variable and is
extracted from the \code{target} field of the \code{Dataset} class.}
}
\description{
This internal function uses an inducer of class \code{InducerGlm}
to fit a Random Forest model to the given training dataset.
}
