% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerRandomForest.R
\name{fit2.InducerRandomForest}
\alias{fit2.InducerRandomForest}
\title{Fit a Random Forest model.}
\usage{
\method{fit2}{InducerRandomForest}(.inducer, task, training_data, response)
}
\arguments{
\item{.inducer}{An S3 object of class \code{InducerRandomForest}.}

\item{task}{String that is either of value \code{"regression"} or \code{"classification"}}

\item{training_data}{Data.Frame or Matrix extracted from the \code{data} field
of the \code{Dataset} class.}

\item{response}{String that stores the name of the target variable and is
extracted from the \code{target} field of the \code{Dataset} class.}
}
\description{
This internal function uses an inducer of class \code{InducerRandomForest}
to fit a Random Forest model to the given training dataset.
}
\details{
We use \code{randomForest::randomForest()} for the model fitting.
}
