% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerXgboost.R
\name{fit2}
\alias{fit2}
\title{Fit an XGBoost model.}
\usage{
fit2(.inducer, task, training_data, response)
}
\arguments{
\item{.inducer}{An S3 object of class \code{InducerXGBoost}.}

\item{task}{String that is either of value \code{"regression"} or \code{"classification"}}

\item{training_data}{Data.Frame or Matrix extracted from the \code{data} field
of the \code{Dataset} class.}

\item{response}{String that stores the name of the target variable and is
extracted from the \code{target} field of the \code{Dataset} class.}
}
\description{
This internal function uses an inducer of class \code{InducerXgboost}
to fit an XGboost model to the given training dataset.
}
\details{
We use \code{xgboost::xgboost()} for the model fitting.
}
