% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparams.R
\name{hp}
\alias{hp}
\title{Hyperparameters S3 class}
\usage{
hp(...)
}
\arguments{
\item{...}{A set of hyperparameters}
}
\description{
A set of functions to define hyperparameters for a model
}
\examples{
hpx <- hp(x = p_num(0, 1), y = p_int(1, Inf), z = p_fct(letters))
hpx
}
