% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparams.R
\name{hpCheck}
\alias{hpCheck}
\title{Hyperparameter check}
\usage{
hpCheck(check, hp)
}
\arguments{
\item{check}{A list of hyperparameters}

\item{hp}{A hyperparameter space}
}
\description{
Check if an element is in the hyperparameter space
}
\examples{
hpx <- hp(x = p_num(0, 1), y = p_int(1, Inf), z = p_fct(letters))
hpCheck(list(x = 1, y = 1, z = "a"), hpx)
}
