% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{hyperparameters}
\alias{hyperparameters}
\title{'hypaerparameters' generic function}
\usage{
hyperparameters(x)
}
\arguments{
\item{x}{An object}
}
\description{
Get information about the hyperparameters of an object
}
\examples{
hyperparameters(splt$cv)
}
