% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{hyperparameters.SplitConstructorCV}
\alias{hyperparameters.SplitConstructorCV}
\title{'hypermarameters' method for SplitCV objects}
\usage{
\method{hyperparameters}{SplitConstructorCV}(x)
}
\arguments{
\item{x}{a SplitCV/SplitConstructorCV object}
}
\description{
Get the hyperparameters information of a Split object
}
\examples{
hyperparameters(splt$cv)
}
