% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{InducerConstructer}
\alias{InducerConstructer}
\title{Inducer S3 class}
\usage{
InducerConstructer(configuration, method)
}
\arguments{
\item{configuration}{Named list of hyperparameters.}

\item{method}{String that stores the name of the method.}
}
\description{
Inducer S3 class
}
