% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerGlm.R
\name{InducerGlm}
\alias{InducerGlm}
\title{Construct or fit a GLM Inducer.}
\usage{
InducerGlm(.data, ...)
}
\arguments{
\item{.data}{Object of class \code{Dataset}. Optional.}

\item{...}{Additional parameters specifying the hyperparameter configuration
of the GLM model. If none are given, the default configuration of
\code{stats::glm} is used for the model fitting process.}
}
\value{
A S3 object of either class \code{Model} (\code{.data} was given) or of class
\code{InducerGlm} (\code{.data} was not given).
}
\description{
Construct or fit a GLM Inducer.
}
\details{
If \code{.data} argument is passed, an Random Forest model will be trained on the given
dataset using the passed hyperparameter configuration. Otherwise, an inducer of
class \code{InducerGlm} is created with the given hyperparameter configuration.
}
