% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerXgboost.R
\name{InducerXGBoost}
\alias{InducerXGBoost}
\title{Construct or fit an XGBoost Inducer.}
\usage{
InducerXGBoost(.data, ...)
}
\arguments{
\item{.data}{Object of class \code{Dataset}. Optional.}

\item{...}{Additional parameters specifying the hyperparameter configuration
of the XGBoost model. If none are given, the default configuration of
\code{xgboost::xgboost} is used for the model fitting process.}
}
\value{
A S3 object of either class \code{Model} (\code{.data} was given) or of class
\code{InducerXGBoost} (\code{.data} was not given).
}
\description{
Construct or fit an XGBoost Inducer.
}
\details{
If \code{.data} argument is passed, an XGBoost model will be trained on the given
dataset using the passed hyperparameter configuration. Otherwise, an inducer of
class \code{InducerXGBoost} is created with the given hyperparameter configuration.
}
\examples{
InducerXGBoost()
InducerXGBoost(verbose = 0)
cars.data <- Dataset(data = cars, target = "dist")
InducerXGBoost(cars.data, verbose = 0, nrounds = 10)

}
