% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{metainfo}
\alias{metainfo}
\title{'metainfo' generic function}
\usage{
metainfo(x)
}
\arguments{
\item{x}{An object}
}
\description{
Get the metainfo of an object
}
\examples{
metainfo(Dataset(data = cars, target = "dist"))
}
