% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr.mini-package}
\alias{mlr.mini}
\alias{mlr.mini-package}
\title{mlr.mini: Mini Package for Machine Learning in R}
\description{
This package is designed to simplify machine learning tasks.
The package offers a unified interface for a variety of machine learning algorithms,
streamlining the modeling process and providing a consistent experience across different models.
Key models, including XGBoost, RandomForest, and general linear models, are incorporated
In addition to model training, the package provides tools for performance evaluation and resampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AdvaProgR-2324/mlr.mini}
  \item Report bugs at \url{https://github.com/AdvaProgR-2324/mlr.mini/issues}
}

}
\author{
\strong{Maintainer}: Sophie H. \email{mlr.sophie@campus.lmu.de}

Authors:
\itemize{
  \item Philip S. \email{mlr.studener@campus.lmu.de}
  \item Pamela T. \email{pamela.tueam@campus.lmu.de}
  \item Sui G. \email{abcd@campus.lmu.de}
  \item Anna N. \email{anna.nazarova@campus.lmu.de}
}

}
