% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelInfo}
\alias{modelInfo}
\title{Model Info}
\usage{
modelInfo(model)
}
\arguments{
\item{model}{A model object}
}
\description{
this function returns the training time in seconds of a model
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
model.xgb <- fit(xgb, cars.data)
modelInfo(model.xgb)
}
