% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparams.R
\name{p_int}
\alias{p_int}
\title{Integer Range}
\usage{
p_int(lower, upper)
}
\arguments{
\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Define ranges for integer hyperparameters
}
\examples{
p_int(1, Inf)
}
