% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparams.R
\name{p_num}
\alias{p_num}
\title{Numeric Range}
\usage{
p_num(lower, upper)
}
\arguments{
\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Define ranges for numeric hyperparameters
}
\examples{
p_num(0, 1)
}
