% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{predict.ModelDummy}
\alias{predict.ModelDummy}
\title{Predictions for Dummy}
\usage{
\method{predict}{ModelDummy}(object, ..., newdata, type = "response")
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}

\item{newdata}{A data frame or Dataset with the new data}

\item{type}{A character string specifying the type of prediction: "response", "se" (standard error), or "prob" (probability). The default is "response".}
}
\description{
this function predicts the majority class in a classification task or the average outcome in a regression task
}
\examples{
# to do
}
