% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{print.EvaluatorAccuracy}
\alias{print.EvaluatorAccuracy}
\title{'print' method for \code{EvaluatorAccuracy}}
\usage{
print.EvaluatorAccuracy(x, ...)
}
\arguments{
\item{x}{An object of class \code{EvaluatorAccuracy}}

\item{...}{Additional arguments}
}
\description{
This method prints the evaluator name and configuration
}
