% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{print.EvaluatorAUC}
\alias{print.EvaluatorAUC}
\title{'print' method for \code{EvaluatorAUC}}
\usage{
\method{print}{EvaluatorAUC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EvaluatorAUC}}

\item{...}{Additional arguments}
}
\description{
This method prints the evaluator name and configuration
}
