% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{print.EvaluatorMAE}
\alias{print.EvaluatorMAE}
\title{'print' method for \code{EvaluatorMAE}}
\usage{
print.EvaluatorMAE(x, ...)
}
\arguments{
\item{x}{An object of class \code{EvaluatorMAE}}

\item{...}{Additional arguments}
}
\description{
This method prints the evaluator name and configuration
}
