% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{print.EvaluatorRMSE}
\alias{print.EvaluatorRMSE}
\title{'print' method for \code{EvaluatorRMSE}}
\usage{
print.EvaluatorRMSE(x, ...)
}
\arguments{
\item{x}{An object of class \code{EvaluatorRMSE}}

\item{...}{Additional arguments}
}
\description{
This method prints the evaluator name and configuration
}
