% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparams.R
\name{print.hp}
\alias{print.hp}
\title{'print' methode for \code{hp} class}
\usage{
print.hp(x, ...)
}
\arguments{
\item{x}{An hp object}

\item{...}{Additional arguments}
}
\description{
Print an hp object
}
\examples{
hpx <- hp(x = p_num(0, 1), y = p_int(1, Inf), z = p_fct(letters))
print(hpx)
}
