% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{print.Inducer}
\alias{print.Inducer}
\title{print method for \code{Inducer} class.}
\usage{
print.Inducer(x, ...)
}
\arguments{
\item{x}{S3 object of class \code{Inducer}.}

\item{...}{Additional arguments.}
}
\description{
print information about inducer.
}
