% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{print.Model}
\alias{print.Model}
\title{print method for \code{Model} class}
\usage{
print.Model(x, ...)
}
\arguments{
\item{x}{S3 object of class \code{Model}.}

\item{...}{Additional arguments.}
}
\description{
print method for \code{Model} class
}
