% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{print.ResamplePrediction}
\alias{print.ResamplePrediction}
\title{'print' method for ResamplePrediction objects}
\usage{
\method{print}{ResamplePrediction}(x, ...)
}
\arguments{
\item{x}{A ResamplePrediction object}

\item{...}{Additional arguments}
}
\description{
Print a ResamplePrediction object
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
cars.split <- splt$cv(folds = 5)(cars.data)
rp <- resample(cars.data, xgb, cars.split)
print(rp)
}
