% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{Split}
\alias{Split}
\title{Resampling S3 Class}
\usage{
Split(data)
}
\arguments{
\item{data}{A Dataset object containing the data}
}
\description{
Split the data into a training and a validation set
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
str(Split(cars.data))
}
