% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{SplitCV}
\alias{SplitCV}
\title{SplitCV method for splitting the data}
\usage{
SplitCV(folds, repeats = 1)
}
\arguments{
\item{folds}{A numeric value: give the number of folds}

\item{repeats}{A numeric value: the number of repetition}
}
\description{
Split the data into 2 parts: the training and validation part
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
cars.data[c(1, 2, 3, 4), ]
cars.data[c(1, 2), "dist"]
}
