% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{splitInstantiate}
\alias{splitInstantiate}
\title{splitInstantiate}
\usage{
splitInstantiate(split, data)
}
\arguments{
\item{split}{A Split object}

\item{data}{A Dataset object containing the data}
}
\description{
Split the data into a training and a validation set
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
cv5 <- splt$cv(folds = 5)
cars.split <- splitInstantiate(cv5, cars.data)
}
