% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\docType{data}
\name{splt}
\alias{splt}
\title{Create an environment for storing Split objects}
\format{
An object of class \code{environment} of length 1.
}
\usage{
splt
}
\description{
Create an environment for storing Split objects
}
\keyword{datasets}
