% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{[.Dataset}
\alias{[.Dataset}
\title{'[' method for subsetting \code{Dataset} objects}
\usage{
\method{[}{Dataset}(x, i, j)
}
\arguments{
\item{x}{A Dataset object}

\item{i}{A vector of row indices}

\item{j}{A vector of column names or indices}
}
\description{
Subset a Dataset object
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
cars.data[c(1, 2, 3, 4), ]
cars.data[c(1, 2), "dist"]
}
