% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InducerXgboost.R
\name{fit2.InducerXGBoost}
\alias{fit2.InducerXGBoost}
\title{Fit an XGBoost model.}
\usage{
\method{fit2}{InducerXGBoost}(.inducer, task, training_data, response)
}
\arguments{
\item{.inducer}{An S3 object of class \code{InducerXGBoost}.}

\item{task}{String that is either of value \code{"regression"} or \code{"classification"}}

\item{training_data}{Data.Frame extracted from the \code{data} field of the \code{Dataset} class.}

\item{response}{String that stores the name of the target variable}
}
\description{
Fit an XGBoost model.
}
