% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{fit2}
\alias{fit2}
\title{S3 (helper) generic for fitting specific Inducers.}
\usage{
fit2(.inducer, task, training_data, response)
}
\arguments{
\item{.inducer}{An S3 object of (sub-)class \code{InducerXxx}.}

\item{task}{String that is either of value \code{"regression"} or \code{"classification"}}

\item{training_data}{Data.Frame or Matrix extracted from the \code{data} field
of the \code{Dataset} class.}

\item{response}{String that stores the name of the target variable and is
extracted from the \code{target} field of the \code{Dataset} class.}
}
\description{
This internal function uses a specific inducer of class \code{InducerXxx}
to fit a corresponding model to the given training dataset.
}
