% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{metainfo.Dataset}
\alias{metainfo.Dataset}
\title{'metainfo' method for \code{Dataset} objects}
\usage{
\method{metainfo}{Dataset}(x)
}
\arguments{
\item{x}{A Dataset object}
}
\description{
Get the metainfo of a Dataset object (name, features, targets, number of observations, type, missing values)
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
metainfo(cars.data)
}
