% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr.mini-package}
\alias{mlr.mini}
\alias{mlr.mini-package}
\title{mlr.mini: Mini Package for Machine Learning in R}
\description{
To do
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AdvaProgR-2324/mlr.mini}
  \item Report bugs at \url{https://github.com/AdvaProgR-2324/mlr.mini/issues}
}

}
\author{
\strong{Maintainer}: Sophie H. \email{mlr.sophie@campus.lmu.de}

Authors:
\itemize{
  \item Philip S. \email{mlr.studener@campus.lmu.de}
  \item Pamela T. \email{pamela.tueam@campus.lmu.de}
  \item Sui G. \email{abcd@campus.lmu.de}
  \item Anna N. \email{anna.nazarova@campus.lmu.de}
}

}
