% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{print.Dataset}
\alias{print.Dataset}
\title{'print' method for Dataset objects}
\usage{
\method{print}{Dataset}(x, ...)
}
\arguments{
\item{x}{A Dataset object}

\item{...}{Additional arguments}
}
\description{
Print a Dataset object
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
print(cars.data)
}
