% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{print.SplitInstance}
\alias{print.SplitInstance}
\title{'print' method for SplitInstance objects}
\usage{
\method{print}{SplitInstance}(x, ...)
}
\arguments{
\item{x}{A SplitInstance object}

\item{...}{Additional arguments}
}
\description{
Print a SplitInstance object
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
cv5 <- splt$cv(folds = 5)
cars.split <- cv5(cars.data)
print(cars.split)
}
