% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{SplitCV}
\alias{SplitCV}
\title{SplitCV method for splitting the data}
\usage{
SplitCV(folds, repeats = 1)
}
\arguments{
\item{folds}{A numeric value: give the number of folds}

\item{repeats}{A numeric value: the number of repetition}
}
\description{
Create a function that splits the data into 2 parts: the training and validation part
}
\examples{
cv3 <- SplitCV(folds = 3, repeats = 2)
cars.data <- Dataset(data = cars, target = "dist")
cars.split <- cv3(cars.data)
}
